Segregate FileProcessor nested types and visitors.
Move ProcessBlockState to file_processor_types.h and visitor classes
(DecoderAllocGuard, DecoderSupportsDispatch, SetDecoderApiCallId,
DispatchVisitor, ProcessVisitor) to file_processor_visitors.h to
reduce the include weight of file_processor.h. Remove vestigial
file_processor.h includes from fps_info.h and vulkan_cpp_consumer_base.h.
