+++ framework/decode/CMakeLists.txt
@@ -81,6 +81,8 @@ target_sources(gfxrecon_decode
                     $<$<BOOL:${D3D12_SUPPORT}>:${CMAKE_CURRENT_LIST_DIR}/dx12_object_info.h>
                     ${CMAKE_CURRENT_LIST_DIR}/file_processor.h
                     ${CMAKE_CURRENT_LIST_DIR}/file_processor.cpp
+                    ${CMAKE_CURRENT_LIST_DIR}/file_processor_types.h
+                    ${CMAKE_CURRENT_LIST_DIR}/file_processor_visitors.h
                     ${CMAKE_CURRENT_LIST_DIR}/preload_file_processor.h
                     ${CMAKE_CURRENT_LIST_DIR}/preload_file_processor.cpp
                     ${CMAKE_CURRENT_LIST_DIR}/file_transformer.h
