+++ framework/decode/file_processor.h
@@ -35,6 +35,7 @@
 #include "util/defines.h"
 #include "decode/decode_allocator.h"
 #include "util/logging.h"
+#include "decode/file_processor_types.h"
 #include "util/file_input_stream.h"
 
 #include <algorithm>
@@ -53,43 +54,10 @@
 GFXRECON_BEGIN_NAMESPACE(gfxrecon)
 GFXRECON_BEGIN_NAMESPACE(decode)
 
-// TODO: Find a better home for the visitors and utilities
-template <bool HasAllocGuard = false>
-struct DecoderAllocGuard
-{};
-
-template <>
-struct DecoderAllocGuard<true>
-{
-    DecoderAllocGuard& operator=(const DecoderAllocGuard&) = delete;
-    DecoderAllocGuard(DecoderAllocGuard&&)                 = delete;
-    DecoderAllocGuard& operator=(DecoderAllocGuard&&)      = delete;
-    DecoderAllocGuard() { DecodeAllocator::Begin(); }
-    ~DecoderAllocGuard() noexcept { DecodeAllocator::End(); }
-};
-
-template <typename Args>
-static bool DecoderSupportsDispatch(ApiDecoder& decoder, const Args& args)
-{
-    if constexpr (DispatchTraits<Args>::kHasCallId)
-    {
-        return decoder.SupportsApiCall(args.call_id);
-    }
-    else if constexpr (DispatchTraits<Args>::kHasMetaDataId)
-    {
-        return decoder.SupportsMetaDataId(args.meta_data_id);
-    }
-    return true;
-}
-
-template <typename Args>
-static void SetDecoderApiCallId(ApiDecoder& decoder, const Args& args)
-{
-    if constexpr (DispatchTraits<Args>::kHasCallId)
-    {
-        decoder.SetCurrentApiCallId(args.call_id);
-    }
-}
+GFXRECON_BEGIN_NAMESPACE(file_processor)
+class DispatchVisitor;
+class ProcessVisitor;
+GFXRECON_END_NAMESPACE(file_processor)
 
 class FileProcessor
 {
@@ -190,16 +158,7 @@ class FileProcessor
 
     void PrintBlockInfo() const;
 
-    enum class ProcessBlockState : int32_t
-    {
-        // Negative values indicate terminal states. Do not call ProcessBlocks again after receiving these.
-        //
-        // Returned when ProcessBlocks ...
-        kFrameBoundary = 1,  // encountered a frame boundary
-        kRunning       = 0,  // never. Internal state: continue looping in ProcessBlocks
-        kEndProcessing = -1, // completed processing (!ContinueDecoding or clean EOF)
-        kError         = -2, // encountered an error
-    };
+    using ProcessBlockState = file_processor::ProcessBlockState;
     static bool ContinueProcessing(ProcessBlockState state) { return static_cast<int32_t>(state) >= 0; }
     static bool IsFrameBoundary(ProcessBlockState state) { return state == ProcessBlockState::kFrameBoundary; }
 
@@ -260,154 +219,12 @@ class FileProcessor
         return *block_parser_;
     }
 
-    class DispatchVisitor
-    {
-      public:
-        // No valid dispatch args, nothing to do. It is possible to modify in future to support passing down
-        // raw block data to some raw block handler if needed
-        void operator()(const std::monostate&) {}
+    using DispatchVisitor = file_processor::DispatchVisitor;
 
-        // Dispatch based on the Args traits.
-        template <typename Args>
-        void operator()(const Args* args)
-        {
-            DispatchArgs(args);
-        }
-
-        // State Marker control
-        void operator()(const StateBeginMarkerArgs* state_begin)
-        {
-            // The block and marker type are implied by the Args type
-            file_processor_.ProcessStateBeginMarker(*state_begin);
-            DispatchArgs(state_begin);
-        }
-
-        void operator()(const StateEndMarkerArgs* state_end)
-        {
-            // The block and marker type are implied by the Args type
-            file_processor_.ProcessStateEndMarker(*state_end);
-            DispatchArgs(state_end);
-        }
-
-        void operator()(const AnnotationArgs* annotation)
-        {
-            if (annotation_handler_)
-            {
-                auto annotation_call = [this](auto&&... expanded_args) {
-                    annotation_handler_->ProcessAnnotation(std::forward<decltype(expanded_args)>(expanded_args)...);
-                };
-                std::apply(annotation_call, annotation->GetTuple());
-            }
-        }
-
-        DispatchVisitor(FileProcessor&                  file_processor,
-                        const std::vector<ApiDecoder*>& decoders,
-                        AnnotationHandler*              annotation_handler) :
-            file_processor_(file_processor),
-            decoders_(decoders), annotation_handler_(annotation_handler)
-        {}
-
-        void SetBlockIndex(uint64_t block_index) { block_index_ = block_index; }
-
-      private:
-        template <typename Args>
-        void DispatchArgs(const Args* args)
-        {
-            constexpr auto decode_method = DispatchTraits<Args>::kDecoderMethod;
-            for (auto decoder : decoders_)
-            {
-                if (DecoderSupportsDispatch(*decoder, *args))
-                {
-                    [[maybe_unused]] DecoderAllocGuard<DispatchTraits<Args>::kHasAllocGuard> alloc_guard{};
-                    SetDecoderApiCallId(*decoder, *args);
-                    decoder->SetCurrentBlockIndex(block_index_);
-                    auto dispatch_call = [&decoder, decode_method](auto&&... expanded_args) {
-                        (decoder->*decode_method)(std::forward<decltype(expanded_args)>(expanded_args)...);
-                    };
-                    std::apply(dispatch_call, args->GetTuple());
-                }
-            }
-        }
-
-        FileProcessor&                  file_processor_;
-        const std::vector<ApiDecoder*>& decoders_;
-        AnnotationHandler*              annotation_handler_;
-        uint64_t                        block_index_;
-    };
+    friend class file_processor::ProcessVisitor;
 
   private:
-    class ProcessVisitor
-    {
-      public:
-        // NOTE: All overloads should set all state, as the caller is *reusing* the Visitor object across a number of
-        //       std::visit calls
-
-        // Frame boundary control
-        void operator()(const FunctionCallArgs* function_call)
-        {
-            is_frame_delimiter = file_processor_.ProcessFrameDelimiter(function_call->call_id);
-            success            = true;
-        }
-
-        void operator()(const MethodCallArgs* method_call)
-        {
-            is_frame_delimiter = file_processor_.ProcessFrameDelimiter(method_call->call_id);
-            success            = true;
-        }
-
-        void operator()(const FrameEndMarkerArgs* end_frame)
-        {
-            // The block and marker type are implied by the Args type
-            is_frame_delimiter = file_processor_.ProcessFrameDelimiter(*end_frame);
-            success            = true;
-        }
-
-        // I/O Control
-        void operator()(const ExecuteBlocksFromFileArgs* execute_blocks)
-        {
-            // The block and marker type are implied by the Args type
-            is_frame_delimiter = false;
-            success            = file_processor_.ProcessExecuteBlocksFromFile(*execute_blocks);
-        }
-
-        void operator()(const StateEndMarkerArgs* state_end)
-        {
-            // The block and marker type are implied by the Args type
-            is_frame_delimiter = false;
-            success            = true;
-            file_processor_.ProcessStateEndMarkerFrameState(*state_end);
-        }
-
-        void operator()(const AnnotationArgs* annotation)
-        {
-            // The block and marker type are implied by the Command type
-            is_frame_delimiter = false;
-            success            = true;
-            file_processor_.ProcessAnnotation(*annotation);
-        }
-
-        void operator()(const std::monostate&) { Reset(); }
-
-        template <typename Args>
-        void operator()(const Args*)
-        {
-            Reset();
-        }
-
-        bool IsSuccess() const { return success; }
-        bool IsFrameDelimiter() const { return is_frame_delimiter; }
-        ProcessVisitor(FileProcessor& file_processor) : file_processor_(file_processor) {}
-        void Reset()
-        {
-            is_frame_delimiter = false;
-            success            = true;
-        }
-
-      private:
-        bool           is_frame_delimiter = false;
-        bool           success            = true;
-        FileProcessor& file_processor_;
-    };
+    using ProcessVisitor = file_processor::ProcessVisitor;
 
     bool ProcessFileHeader();
