+++ framework/decode/file_processor_types.h
@@ -1,0 +1,50 @@
+/*
+** Copyright (c) 2018-2026 LunarG, Inc.
+**
+** Permission is hereby granted, free of charge, to any person obtaining a
+** copy of this software and associated documentation files (the "Software"),
+** to deal in the Software without restriction, including without limitation
+** the rights to use, copy, modify, merge, publish, distribute, sublicense,
+** and/or sell copies of the Software, and to permit persons to whom the
+** Software is furnished to do so, subject to the following conditions:
+**
+** The above copyright notice and this permission notice shall be included in
+** all copies or substantial portions of the Software.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+** IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+** FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+** AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+** LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING
+** FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
+** DEALINGS IN THE SOFTWARE.
+*/
+
+#ifndef GFXRECON_DECODE_FILE_PROCESSOR_TYPES_H
+#define GFXRECON_DECODE_FILE_PROCESSOR_TYPES_H
+
+#include "util/defines.h"
+
+#include <cstdint>
+
+GFXRECON_BEGIN_NAMESPACE(gfxrecon)
+GFXRECON_BEGIN_NAMESPACE(decode)
+GFXRECON_BEGIN_NAMESPACE(file_processor)
+
+// ProcessBlocks specific status.
+enum class ProcessBlockState : int32_t
+{
+    // Negative values indicate terminal states. Do not call ProcessBlocks again after receiving these.
+    //
+    // Returned when ProcessBlocks ...
+    kFrameBoundary = 1,  // encountered a frame boundary
+    kRunning       = 0,  // never. Internal state: continue looping in ProcessBlocks
+    kEndProcessing = -1, // completed processing (!ContinueDecoding or clean EOF)
+    kError         = -2, // encountered an error
+};
+
+GFXRECON_END_NAMESPACE(file_processor)
+GFXRECON_END_NAMESPACE(decode)
+GFXRECON_END_NAMESPACE(gfxrecon)
+
+#endif // GFXRECON_DECODE_FILE_PROCESSOR_TYPES_H
